Tier one: tangents, normal maps, mip chains, 32-bit indices, emission
Four gaps that all touch the vertex format and the material, so they go in
together.

TANGENTS. VertexIn gains a vec4: xyz the direction u increases in, w the
bitangent's handedness. GenerateTangents solves the per-triangle 2x2 uv
system (Lengyel), accumulates, orthogonalises against the normal and
records the sign. Handedness is not optional -- a mirrored uv shell is half
of every character model, and without it every dent on that half is a bump.

The generator is called from each mesh maker AND from UploadMesh, which
regenerates when the first vertex's tangent is zero. One choke point that
cannot be forgotten: a normal map on a mesh with a zero tangent does not
look wrong, it makes the TBN singular and the surface NaN-black. glTF
TANGENT is read when the file has it, because an exporter's tangents are
the ones the map was baked against.

NORMAL / METALLIC / EMISSIVE / OCCLUSION maps. Only xy of the normal map
are read and z is rebuilt, so a two-channel map works unchanged and an
8-bit three-channel one is more accurate than its stored blue. Emission is
RADIANCE, not a 0..1 colour -- the scene target is half-float precisely so
a filament can be forty times a wall, and the bloom threshold is in the
same units. Baked occlusion multiplies the ambient and image-based terms
only: a crevice the sky cannot reach is still lit by a lamp aimed into it.

MIP CHAINS, sRGB and 16x ANISOTROPY. CreateTexture2D builds and fills the
chain by default. Not "slightly softer at distance" -- the floor in the
test went from an adjacent-pixel energy of 7720 to 377, and TAA cannot fix
that class of aliasing because it averages samples of an already
undersampled signal. Anisotropy is the other half: in the band where the
squeeze is between 2:1 and 16:1 it keeps 14.0 of contrast where an
isotropic mip keeps 8.1. It is a setting rather than a constant, partly
because it is a quality trade and partly because a hard-coded value cannot
be tested.

32-BIT INDICES throughout. The glTF importer used to refuse anything over
65535 vertices outright and the terrain silently returned an empty chunk.

Also, the shadow map gets a sampler of its own. Slot 0 was filtering the
material maps AND the shadow map, so it could not be Repeat for tiling
albedo and Clamp for a shadow lookup at once -- which is what blocked the
fullscreen clamp fix last time. Now the material sampler wraps at 16x, the
fullscreen one clamps, and shadows have their own.

Measured, in apps/surface and mesh_test:

  a flat normal map differs from no map by 1/255
  a normal-map tilt of 0.45 rad shades at 139.0; the same tilt applied to
    the GEOMETRY shades at 139.0, and no tilt at 180.0
  normal_strength 0 reproduces "no map" exactly, then 165 / 139 / 61
  a mirrored uv shell renders identically under a bitangent-direction tilt,
    rms 0.00, which is what the handedness sign is for
  sphere tangents are perpendicular to 6.7e-08 and unit to 1.2e-07
  aliasing falls to 4.9% with mips, and the mean rises rather than falls
  a 90601-vertex sphere covers the same silhouette as a 24x32 one to 0.4%
  emission is monotonic across 0.3 to 22 and keeps its colour ratio

Eight mutations, all caught: the bitangent sign ignored in the shader and
at mesh time, z not rebuilt, Gram-Schmidt dropped, strength scaling the
vector instead of the slope, mip generation skipped, the sRGB flag ignored,
anisotropy forced to 1.

+++ apps/skinned/offscreen.cpp
@@ -688,15 +688,15 @@ int main() {
                 // origin and very easy to blame on the exporter.
                 for (int c = 0; c < 4; ++c) { sv.joints[c] = 0; sv.weights[c] = 0.0f; }
             } else {
-                sv.joints[0] = std::uint16_t(i % 2);
+                sv.joints[0] = std::uint32_t(i % 2);
                 sv.weights[0] = 1.0f;
             }
             ssk.push_back(sv);
         }
         for (int i = 0; i + 2 < kSmall; ++i) {
-            sm.indices.push_back(std::uint16_t(i));
-            sm.indices.push_back(std::uint16_t(i + 1));
-            sm.indices.push_back(std::uint16_t(i + 2));
+            sm.indices.push_back(std::uint32_t(i));
+            sm.indices.push_back(std::uint32_t(i + 1));
+            sm.indices.push_back(std::uint32_t(i + 2));
         }
         const eng::MeshHandle small = renderer->UploadSkinnedMesh(sm, ssk, 2);
         Check(Valid(small), "a small skinned mesh uploaded");
